support mnist test and softmax in vm32

+++ vm.c
@@ -75,7 +75,7 @@ int vm_read_weights(uint8_t *host_addr, int len)
 
 // TODO: 当weight数据不够时，处理异常
 int vm_maxtrix_multiply(uint32_t unified_buffer_addr, uint16_t accumulator_addr,
-        uint8_t input_row, uint8_t input_col, uint8_t weight_row, uint8_t weight_col)
+        uint16_t input_row, uint16_t input_col, uint16_t weight_row, uint16_t weight_col)
 {
     if (weight_fifo.size < weight_row * weight_col) {
         DBG("vm32_maxtrix_multiply failed, weight fifo is too small\n");
