+++ vm.h
@@ -15,7 +15,7 @@ int vm_read_host_memory(uint32_t unified_buffer_addr, uint8_t *host_addr, int le
 int vm_read_weights(uint8_t *host_addr, int len);
 
 int vm_maxtrix_multiply(uint32_t unified_buffer_addr, uint16_t accumulator_addr,
-        uint8_t input_row, uint8_t input_col, uint8_t weight_row, uint8_t weight_col);
+        uint16_t input_row, uint16_t input_col, uint16_t weight_row, uint16_t weight_col);
 
 int vm_activate(act_type_enum_t type, uint16_t accumulator_addr, uint32_t unified_buffer_addr, int len);
